First Upload
This is partly built. The pipeline is completed.

There needs to be two controllers for the instructions and one for the load and stores that need to be completed.

The cache memory is almost done as well.

Then the GTE from laxer3a is almost completed, but the FSM for the instructions need to be tested with it.

Some of the COP0 needs to be done is the breakpoint systems